perf(test): split ApiCheckModalHost test into 4 subject files
The 4962-line file (84 tests) was the largest single-file tail in CI
(~121s wall time on one worker). Split into modal/extraction/models/
probesAndProfiles files with shared mock and support modules so vitest
can run them on parallel workers within a shard. No test bodies changed.
